pyproject.toml
[project]
name = "modoo-marble-backend"
version = "0.1.0"
requires-python = ">=3.11,<3.14"
dependencies = [
  "fastapi[standard] (>=0.135.1,<0.136.0)",
  "uvicorn[standard]>=0.30,<1.0.0",
  "gunicorn>=23.0.0,<26.0.0",
  "python-socketio[client]>=5.16,<6.0.0",
  "tortoise-orm>=0.20,<0.21",
  "aerich>=0.7.2,<0.8.0",
  "asyncpg>=0.30,<1.0.0",
  "redis>=5.0,<8.0.0",
  "apscheduler>=3.10,<4.0.0",
  "google-genai>=0.3,<2.0.0",
  "tenacity>=8.2,<10.0.0",
  "pybreaker>=1.0,<2.0.0",
  "python-jose[cryptography]>=3.3,<4.0.0",
  "passlib[bcrypt]>=1.7,<2.0.0",
  "pydantic>=2.6,<3.0.0",
  "pydantic-settings>=2.2,<3.0.0",
  "python-dotenv>=1.0,<2.0.0",
  "structlog>=24.1,<26.0.0",
  "pyjwt>=2.11.0",
]

[dependency-groups]
dev = [
  "ruff>=0.6,<1.0.0",
  "pytest>=8.2,<9.0.0",
  "pytest-asyncio>=0.23,<0.24.0",
  "pytest-mock>=3.14,<4.0.0",
  "httpx>=0.27,<1.0.0",
  "pre-commit>=4.5.1",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "ASYNC"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.aerich]
tortoise_orm = "app.config.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."
